refactor: extract the application subproject
