+++ .idea/compiler.xml
@@ -29,6 +29,8 @@
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.eclipse.jgit/org.eclipse.jgit/4.4.1.201607150455-r/63998ced66e425d9e8bcd0c59f710c98f0c021ff/org.eclipse.jgit-4.4.1.201607150455-r.jar" />
         </processorPath>
         <module name="gradle-boilerplate.test" />
+        <module name="gradle-boilerplate.domain.test" />
+        <module name="gradle-boilerplate.application.test" />
       </profile>
       <profile name="Gradle Imported" enabled="true">
         <outputRelativeToContentRoot value="true" />
@@ -58,6 +60,8 @@
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.eclipse.jgit/org.eclipse.jgit/4.4.1.201607150455-r/63998ced66e425d9e8bcd0c59f710c98f0c021ff/org.eclipse.jgit-4.4.1.201607150455-r.jar" />
         </processorPath>
         <module name="gradle-boilerplate.main" />
+        <module name="gradle-boilerplate.domain.main" />
+        <module name="gradle-boilerplate.application.main" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel target="11" />
