+++ application/build.gradle
@@ -1,0 +1,15 @@
+apply plugin: 'java-library'
+apply plugin: 'io.spring.dependency-management'
+
+dependencyManagement {
+    imports {
+        mavenBom ext.mavenBom
+    }
+}
+
+dependencies {
+    annotationProcessor "org.immutables:value:${immutablesVersion}"
+    compileOnly 'com.google.code.findbugs:jsr305:3.0.2'
+    compileOnly 'org.springframework:spring-context'
+    implementation project(':domain')
+}
