+++ build.gradle
@@ -26,6 +26,7 @@ allprojects {
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation project(':domain')
+    implementation project(':application')
     compileOnly 'com.google.code.findbugs:jsr305:3.0.2'
     compileOnly "org.immutables:value:${immutablesVersion}:annotations"
     annotationProcessor "org.immutables:value:${immutablesVersion}"
