+++ settings.gradle
@@ -1,1 +1,2 @@
 include ':domain'
+include ':application'
